+++ include/linux/nvhost.h
@@ -727,7 +727,7 @@ const char *nvhost_syncpt_get_name(struct platform_device *dev, int id);
 void nvhost_syncpt_cpu_incr_ext(struct platform_device *dev, u32 id);
 int nvhost_syncpt_read_ext_check(struct platform_device *dev, u32 id, u32 *val);
 int nvhost_syncpt_wait_timeout_ext(struct platform_device *dev, u32 id, u32 thresh,
-	u32 timeout, u32 *value, struct timespec64 *ts);
+	unsigned long timeout, u32 *value, struct timespec64 *ts);
 int nvhost_syncpt_create_fence_single_ext(struct platform_device *dev,
 	u32 id, u32 thresh, const char *name, int *fence_fd);
 void nvhost_syncpt_set_min_eq_max_ext(struct platform_device *dev, u32 id);
